+++ src/Type/ProductLookupBy.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * This file is part of the Billbee API package.
+ *
+ * Copyright 2017 - 2018 by Billbee GmbH
+ *
+ * For the full copyright and license information, please read the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Created by Julian Finkler <julian@billbee.io>
+ */
+
+namespace BillbeeDe\BillbeeAPI\Type;
+
+class ProductLookupBy
+{
+    const ID = 'id';
+    const EAN = 'ean';
+    const SKU = 'sku';
+}